refactor: extract shared buildMatcher into src/rules/match-paths.ts
filename-casing, forbid-content-pattern, and enforce-tdd each had
a near-identical gitignore-style picomatch shim. Promote the shared
buildMatcher to its own module so the three rules depend on one
matcher implementation and new rules can reuse it.
